requirements.txt
pandas~=2.0.1
matplotlib~=3.10.1
seaborn~=0.13.2
Unidecode~=1.3.8
numpy~=1.24.3
cryptography~=41.0.0
colorama~=0.4.6
geopandas~=0.14.3
imageio~=2.37.0
geopy~=2.4.1
plotly~=5.7.0
dash~=2.3.1
pip~=22.3
attrs~=23.1.0
protobuf~=5.29.4
pillow~=9.5.0
tornado~=6.3.2
Jinja2~=3.1.2
setuptools~=65.5.0
packaging~=23.1
pyparsing~=3.0.9
typing_extensions~=4.13.1
future~=1.0.0
streamlit~=1.44.1
openpyxl~=3.1.1
GitPython~=3.1.44
gitdb~=4.0.12
cffi~=1.15.1
PySocks~=1.7.1
Flask~=2.3.2
Flask-Compress~=1.17
selenium~=4.14.0
requests~=2.31.0
PyYAML~=6.0
pytz~=2023.3
toml~=0.10.2
click~=8.1.3
fiona~=1.10.1
cligj~=0.7.2
shapely~=2.1.0
click-plugins~=1.1.1
MarkupSafe~=2.1.2
Werkzeug~=2.3.6
blinker~=1.6.2
itsdangerous~=2.1.2
requests-oauthlib~=1.3.1
aiohttp~=3.8.4
yarl~=1.9.2
geographiclib~=2.0
smmap~=5.0.2
charset-normalizer~=3.1.0
altair~=5.5.0
jsonschema~=4.23.0
narwhals~=1.33.0
pyarrow~=19.0.1
referencing~=0.36.2
traitlets~=5.9.0
python-dateutil~=2.8.2
six~=1.16.0
tenacity~=9.1.2
pydeck~=0.9.1
pyproj~=3.7.1
certifi~=2023.5.7
tzdata~=2023.3
cycler~=0.12.1
Pygments~=2.15.1
zstandard~=0.23.0
urllib3~=1.26.16
et-xmlfile~=1.1.0
idna~=3.4
watchdog~=6.0.0
contourpy~=1.0.7
bokeh~=3.1.1
fonttools~=4.39.4
pycparser~=2.21
cachetools~=5.3.1
jsonschema-specifications~=2024.10.1
kiwisolver~=1.4.4
openrouteservice~=2.3.3
scipy~=1.15.3
googlemaps~=4.10.0
python-dotenv~=1.1.0